requirements.txt
packaging
pyyaml
flask
redis
